reintroduce utilities.py and exceptions.py for lite text gen
